# openapi.yaml
openapi: 3.0.0
info:
  title: Timesheet API
  description: API for managing employee clock-in and clock-out records
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Development server

paths:
  /api/clock-in:
    post:
      summary: Clock in employee
      description: Record employee clock-in time
      tags:
        - Clock Records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: Employee email address
      responses:
        "200":
          description: Clock-in recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      clock_in_time:
                        type: string
                        format: date-time
                      created_at:
                        type: string
                        format: date-time
        "399":
          description: Invalid request
        "401":
          description: Unauthorized

  /api/clock-out:
    post:
      summary: Clock out employee
      description: Record employee clock-out time
      tags:
        - Clock Records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: Employee email address
      responses:
        "200":
          description: Clock-out recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      clock_in_time:
                        type: string
                        format: date-time
                      clock_out_time:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: No active clock-in record found

  /api/history/{email}:
    get:
      summary: Get employee clocking history
      description: Retrieve clock-in and clock-out records for an employee
      tags:
        - History
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
            format: email
          description: Employee email address
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of records to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of records to skip
      responses:
        "200":
          description: Clocking history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        email:
                          type: string
                        clock_in_time:
                          type: string
                          format: date-time
                        clock_out_time:
                          type: string
                          format: date-time
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
        "400":
          description: Invalid request
        "404":
          description: Employee not found

  /health:
    get:
      summary: Check API health
      description: Verify that the API is running
      tags:
        - Health
      responses:
        "200":
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT